requirements.txt
# Environment & Config
python-dotenv
 
# PDF + OCR Processing
pymupdf
pdf2image
pytesseract
pillow
 
# Data Handling
pandas
numpy
 
# Database
pymongo
 
# Embeddings
langchain
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
 
# LLM (Groq)
langchain-groq
 
# Text Splitting (used in RAG pipelines)
langchain-text-splitters
 
# Streamlit UI
streamlit